Moved convenience methods to a separate file
